Add processQueues method to ObjectManager and update rendering logic
